LayoutTests/fast/exclusions/wrap-through-parsing-expected.txt

Test parsing of the CSS wrap-through property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test("-webkit-wrap-through: wrap") is "wrap"
PASS test("-webkit-wrap-through: none") is "none"
PASS test("-webkit-wrap-through: ;") is ""
PASS test("-webkit-wrap-through: 5") is ""
PASS test("-webkit-wrap-through: -1.2") is ""
PASS test("-webkit-wrap-through: 'string'") is ""
PASS testComputedStyle("wrap") is "wrap"
PASS testComputedStyle("5") is "wrap"
PASS testComputedStyle("'string'") is "wrap"
PASS testNotInherited("wrap", "none") is "none"
PASS testNotInherited("none", "wrap") is "wrap"
PASS successfullyParsed is true

TEST COMPLETE
